ai-backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0

# LangChain stack
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0

# Vector DB
chromadb>=0.5.0

# OpenAI (ya no se usa como fallback, mantenemos por si acaso)
# openai>=1.10.0

# Groq (cloud fallback — gratuito y ultra-rápido)
groq>=0.9.0

# Document ingestion
pdfplumber>=0.11.0

# Audio/Video transcription (requires ffmpeg installed separately)
faster-whisper>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
